.github: drop testing job and reduce building job
We don't need testing jobs at the current stage, as not all geth tests
are healthy and we're not going to fix them. We also don't need linter job, as
we're not going to fix linter problems within the whole project to make it
possible to merge changes from the geth repo without conflicts.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>