dbft: rework NewBlockFromContext and ProcessBlock callbacks
NewBlockFromContext must operate with a pure context and PrepareRequest,
in particular. It must not include transactions, because some of the
transactions may be missing by the moment of call to NewBlockFromContext.
Close #31.

dBFT-consistent Block structure should not include types.Block itself,
it's sufficient to have types.Header and a set of transactions.

ProcessBlock callback should create new block from the proposed header
by adding signature to the header Extra and by adding transactions to the
resulting types.Block.

Avoid creating a lot of block copies, avoid usage of the whole Block where
it's sufficient to have only Header.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>