dbft: include proposed parent witness verification to VerifyPrepareRequest callback
We need to verify the proposed parent's witness in VerifyPrepareRequest, otherwise
the proposed parent may be invalid. Other PrepareRequest fields verification will
be done within the scope of #53.

Also, properly reinitialize sealing proposal on backup nodes after PrepareRequest
verification. This scheme won't work with non-empty blocks after the first CV, but we
have #54 for that.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>