dbft: directly insert newly added block into chain
The way of insertion is OK, the only thing that may be optimized may be
done after dBFT context initialization refactoring (ref. #53 and #60). Once
we have verified set of receipts in VerifyPrepareRequest callback, we may
store this set of receipts and insert block in chain with the known and
calculated receipts.

This optimisation will be discussed in a separate issue, ref. #60.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>