miner: do not skip empty block proposal submitting for dBFT
A part of #54. Proposal reinitialisation should happen frequently irrespectively
to the fact of new transactions arrival. This commit makes miner to propose
new sealing task with the frequency conigured via --miner.recommit node setting
instead of producing exactly one task per height if no transactions are added
to the pool. This modification is needed to fetch up-to-date block information
after dBFT ChangeViews so that new primary is able to build PrepareRequest with
fresh data.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>