dbft: do not return error on new seal proposal if sealing
Miner can submit seal proposals as frequently as it configured, and it's not
an error to submit a fresh proposal when the old one is still sealing.
This change leads to the fact that one sealing task is always present in
the miner's sealing pool and stale tasks are rejected from the miner's
POW, but DBFT service itself is able to always receive fresh tasks.

It's kind of miner's misuse, but it does not lead to unwanted consequences.
This implementation will be kept until refactoring is done in a separate task,
ref. #71.

A part of #54.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>