dbft: make header's Coinbase field be configurable
Make header's Coinbase be configurable via miner.etherbase node setting.
This commit allows Coinbase be different from the validator's address to be
able to send reward to a side account, e.g. to some Governance contract.
Validator-related coinbase check is removed as obsolete. This change is
needed to support NeoX governance model.

It is guaranteed by miner's constructor for Coinbase not to be empty. No
other validity checks is performed.

A part of #86.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>