dbft: use precalculated receipts (if available) to store block
The idea of this commit is to keep block's receipts got after
PrepareRequest level block verification and use them while submitting
the resulting block to the chain in order to save some time on receipts
recalculation. Close #60.

Originally this commit aimed to solve the problem described in #100.
However, seems that direct receipts-less block insertion also works
fine and it's not the root of the problem described in #100.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>