dbft: do not call postBlock callback immediately in ProcessBlock
There's a chance that newly-accepted block will be added by worker using
the corresponding sealing task saved in the worker state. If so, then
BlockQueue don't have any way to track the end of block addition
process. Since postBlock contains NextConsensus computations, it must be
called only after successfull block persist so that the proper state is
available in the chain. And thus, call postBlock only when chain
notification arrives.

A part of #14. An example of problem logs that this commit is aimed to
fix:
```
2024-03-11T19:52:22.324+0300	DEBUG	dbft@v0.0.0-20240130131324-6d59675ad370/dbft.go:214	received message	{"type": "Commit", "from": 5, "height": 1, "view": 1, "my_height": 1, "my_view": 1}
2024-03-11T19:52:22.324+0300	DEBUG	dbft@v0.0.0-20240130131324-6d59675ad370/dbft.go:214	received message	{"type": "Commit", "from": 6, "height": 1, "view": 1, "my_height": 1, "my_view": 1}
2024-03-11T19:52:22.324+0300	INFO	dbft@v0.0.0-20240130131324-6d59675ad370/dbft.go:511	received Commit	{"validator": 6}
2024-03-11T19:52:22.325+0300	INFO	dbft@v0.0.0-20240130131324-6d59675ad370/check.go:66	approving block	{"height": 1, "hash": "8633939a9fa52e0fc260b8dc335f21bf11c5519ea150683d434e4986f597a294", "tx_count": 0, "merkle": "eeccdc5021307b4b06231fcd6c7b7a9c632863d647459bb4124e9c95f26e50f5", "prev": "9ee34b625f2e9aac3d0b7aaaea40db845906d6d24c89e3e4f53af8d63074eb83"}
INFO [03-11|19:52:22.325] calculating validators                   height=1
CRIT [03-11|19:52:22.325] failed to compute next block validators  err="header not found"
```

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>