dbft: always recalculate block state got from miner
We need an access to block state at the PrepareRequest construction
level to be able to calculate NextConsensus for the current proposal.
State for the 0-th view can be fetched from miner, but I'd avoid adding
one more miner-dependent state field to the dBFT engine, especially
given the fact that one day we'll remove state calculations from
miner.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>