dbft: cache validators computed based on the accepted state
If validators are requested by block height then the block state is
already in chain and accepted, and thus, the set of validators
computed based on this state won't change.

Cache validators for three most recent heights which is enough for
effective network payloads processing (we may have stale payloads
travelling through the network) and dBFT fuctioning (it always needs
the latest set of validators for block signing).

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>