dbft: use block's header to process new blocks from chain
The whole block is not required for this and it's better to explicitly
specify that header is required since headers chain may be ahead of the
blocks chain.

Technically, this commit makes the code slower a bit due to headers
copying, but functionally there's no changes and this commit is required
for proper Downloader event handlers.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>