params: start v0.2.2 cycle
Start v0.2.2 that is aimed to be compatible with v0.2.1 released for
MainNet launch. If there are some urgent fixes need to be applied at the
top of v0.2.1 for MainNet launch, we can add them into this release. If
not, then the next v0.3.0 cycle will be started for major anti-MEV
protocol changes.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>