dbft: untie BlockQueue from miner
Don't use miner for block insertion anymore, dBFT has everything to make
it by the means of BlockQueue. Close #297.

Sometimes it may happen that the node receives new block from the network
earlier than it comes from the dBFT. In this case miner don't cancel
sealing task send to dBFT, and thus, BlockQueue is trying to send result
back to miner. Miner doesn't have the corresponding task (because all
tasks are already cleared from the list), and hence we've got an error
in logs.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>