dbft: take care of Primary's PreBlock's state
Primary node doesn't perform verification of PreBlock because it's
constructed by the node itself. Hence, cachable PreBlock fields are
missing from the Primary's PreBlock context. This commit fixes it and
makes sure that Primary is able to construct the final block based on
the PreBlock's state in case if there are no Envelopes in the proposed
block.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>