contracts: move encrypted tx gas limit check to consensus level
This commit solves two problems:
1. By the moment of GovReward contract execution we must be sure that
   encrypted transaction gas limit satisfies minimum required value.
2. By the same moment we must be sure that Envelope transaction has
   enough gas limit to pay for execution of decrypted transaction.

If these two constrains are violated, contract itself can't do anything
about it, the Envelope execution will be aborted, but still, Envelope
will be accepted to the chain since we don't filter out failed
Envelopes (and we shouldn't do that, ref.
https://github.com/bane-labs/go-ethereum/pull/422#discussion_r1971545809).

However, the check is still needed because otherwise it's easy for the
user to accidently send an invalid Envelope and it will be accepted to
the chain spending user's funds for nothing. I won't say that it's a
vector of attack because small Envelope fee will lead to the fact that
encrypted tx execution itself will be restricted by this value and hence
it's not critical for the system and doesn't bring any profit to the
user.

This commit moves these two checks to mempool and state transition level
so that it's impossible to submit Envelope that is improperly
constructed.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>