dbft: prohibit to initialize local pool with empty root (#476)
For the case of dBFT it's not correct to initialize local pool with
empty root because it's critical for dBFT to have relevant state for
proper proposal verification. dBFT must guarantee that by the moment of
static pool initialization the node's state at the given root is
present; if it's not true then it's a bug that should be fixed on the
dBFT side.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>